Add files via upload
First attempt to make a user interface to machine.
When using Involt to connect Arduino to an NW application, an interface for the Machine was generated. This is because the machine will probably connect via Arduino to the computer or other device.
It is important to understand the environment of the involt framework before the first attempt to modify the interface. Involt provides a communication approach mainly between an application and an arduino. The versatility of NW allows the project to expand to multiple platforms without worrying about the hardware that contains the application.

We started to try 3 things:
One: receive an analog input signal and display it in the interface
Two: send a 0-255 int to pwm Output from the interface to the machine
Three: Send a binary signal to arduino.

One: it is for the input of a sensor and the processing of this information.
Two: it is to emulate a dispensing of material from the interface to the machine.
Three: simulate an open door or close from the interface to the machine.

for Three we use a signal sent to pin 13.
For Dos we use 3 leds connected to different pins (11, 10 and 9). We send a live signal from the interface. The quantity could be visualized from the light intensity of the LEDs.
For One we connect a potentiometer to vary the Signal Strength, and we show it on the screen.

These examples are Guided in the involt page in github. Are modifications of: 
Blink tutorial
2 Knobs 1 App
"Do a barrel roll"